+++ src/world/ObjectMgr.cpp
@@ -143,7 +143,7 @@ ObjectMgr::~ObjectMgr()
     }
 
     Log.Notice("ObjectMgr", "Deleting Charters...");
-    for (int i = 0; i < NUM_CHARTER_TYPES; ++i)
+    for (uint8 i = 0; i < NUM_CHARTER_TYPES; ++i)
     {
         for (std::unordered_map<uint32, Charter*>::iterator itr = m_charters[i].begin(); itr != m_charters[i].end(); ++itr)
         {
@@ -2036,7 +2036,7 @@ void ObjectMgr::LoadTrainers()
                     ts.pCastSpell = dbcSpell.LookupEntryForced(CastSpellID);
                     if (ts.pCastSpell)
                     {
-                        for (int k = 0; k < 3; ++k)
+                        for (uint8 k = 0; k < 3; ++k)
                         {
                             if (ts.pCastSpell->Effect[k] == SPELL_EFFECT_LEARN_SPELL)
                             {
@@ -2987,7 +2987,7 @@ void Charter::SaveToDB()
 Charter* ObjectMgr::GetCharterByItemGuid(uint64 guid)
 {
     m_charterLock.AcquireReadLock();
-    for (int i = 0; i < NUM_CHARTER_TYPES; ++i)
+    for (uint8 i = 0; i < NUM_CHARTER_TYPES; ++i)
     {
         std::unordered_map<uint32, Charter*>::iterator itr = m_charters[i].begin();
         for (; itr != m_charters[i].end(); ++itr)
