+++ src/world/TerrainMgr.cpp
@@ -25,16 +25,16 @@
 
 TerrainHolder::TerrainHolder(uint32 mapid)
 {
-    for (int32 i = 0; i < TERRAIN_NUM_TILES; ++i)
-        for (int32 j = 0; j < TERRAIN_NUM_TILES; ++j)
+    for (uint8 i = 0; i < TERRAIN_NUM_TILES; ++i)
+        for (uint8 j = 0; j < TERRAIN_NUM_TILES; ++j)
             m_tiles[i][j] = NULL;
     m_mapid = mapid;
 }
 
 TerrainHolder::~TerrainHolder()
 {
-    for (int32 i = 0; i < TERRAIN_NUM_TILES; ++i)
-        for (int32 j = 0; j < TERRAIN_NUM_TILES; ++j)
+    for (uint8 i = 0; i < TERRAIN_NUM_TILES; ++i)
+        for (uint8 j = 0; j < TERRAIN_NUM_TILES; ++j)
             UnloadTile(i, j);
 }
